requirements.txt
# Core dependencies
pydantic>=2.10.6
python-dotenv>=1.0.1

# LangChain and LLM dependencies - latest versions
langchain==0.3.19
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-anthropic==0.3.8
openai>=1.64.0
anthropic>=0.47.2

# LangGraph - latest version
langgraph==0.2.74

# PDF processing
PyMuPDF>=1.25.3
Pillow>=11.1.0
pypdf>=5.3.0

# Vector stores
langchain-qdrant>=0.2.0
qdrant-client>=1.13.2

# Logging
rich>=13.9.4

# Type hints compatibility
typing-extensions>=4.12.2